# Conflicts:
#	src/main/java/gg/projecteden/nexus/features/events/Events.java
#	src/main/java/gg/projecteden/nexus/features/events/y2020/pride20/Pride20Command.java
#	src/main/java/gg/projecteden/nexus/features/listeners/Sleep.java
#	src/main/java/gg/projecteden/nexus/features/mobheads/MobHeads.java

+++ src/main/java/gg/projecteden/nexus/features/minigames/models/matchdata/PixelPaintersMatchData.java
@@ -32,7 +32,7 @@ public class PixelPaintersMatchData extends MatchData {
 	private long roundStart;
 	private int totalFinished;
 	private int roundCountdownId;
-	private int timeLeft;
+	private long timeLeft;
 	private boolean animateLobby;
 	private int lobbyDesign;
 	private int animateLobbyId;
